Rename files after main class
